Refactor code and document examples
The changes involve code refactoring and the addition of documentation cells for better readability in the script. A minor modification also checks for a None type in the `_array` function within the `backend_tensor.py` file. In addition, the Python file `01_gravity_gradiometry_inversion_pgi.py` has been renamed and moved to the temp directory.