[BUG][ENH] Fixed index bug and enhanced block data structures
This commit primarily resolves an indexing bug in interp_output.py that affected the block data structures. Also, it improves the handle of dense grid data in various methods by clearly specifying the dense grid instead of the octree grid. Several test cases were updated for compatibility with the changes.